button: don't sleep in hrtimer callback
Access to the BBAPI code in BIOS is protected by a mutex. bbapi_button will
continuously acquire and release this mutex. Previously this was in IRQ,
context, so whenever another BBAPI call such as a write to the CX2100 display
already locked the mutex, this would lead to:
"BUG: scheduling while atomic"
With a callstack such as:
[  230.383590] Call Trace:
[  230.383595]  <IRQ>
[  230.383603]  dump_stack+0x66/0x8b
[  230.383608]  __schedule_bug+0x51/0x70
[  230.383612]  __schedule+0x625/0x6d0
[  230.383617]  ? update_load_avg+0x615/0x640
[  230.383621]  schedule+0x2f/0xa0
[  230.383624]  schedule_preempt_disabled+0xa/0x10
[  230.383627]  __mutex_lock.isra.9+0x46a/0x4e0
[  230.383631]  ? enqueue_task_fair+0xa6/0x490
[bbapi]
[bbapi]
[bbapi_button]
[bbapi]
[bbapi_button]
[  230.383651]  __hrtimer_run_queues+0x108/0x290
[  230.383654]  hrtimer_interrupt+0xe5/0x240
[  230.383658]  smp_apic_timer_interrupt+0x6a/0x130
[  230.383661]  apic_timer_interrupt+0xf/0x20
[  230.383663]  </IRQ>

With this patch we postpone the wait for mutex from hrtimer callback into a
kernel worker thread.