fructose: silence some warnings
Using get_test_name() and drop the test_case_name function parameter in all
the functions might be the correct solution. However, we just fix this one
occurance to match the others and silence the compiler complaining about
unused test_case_name parameter.

The other fix is related. "msg1" and "msg2" function parameter is not used in
"test_root::test_assert_same_data" so we make them anonymous to silence the
compiler warnings.