debian: precompiled kernel module with correct maintainer and description
By default, dkms mkbmdeb uses a template stored under the path. Based on
this template, dkms mkbmdeb generates a debian package. The resulting
package does not contain the maintainer or the detailed description of
BBAPI. However, it is desirable that the generated package contains
the above information. Therefore, in this commit the template is adapted
so that the appropriate maintainer and package description of BBAPI are
included in the package.

Reviewed-by: Patrick Bruenn <p.bruenn@beckhoff.com>