api: expose search area as module parameter
If running inside a hypervisor like Hyper-V, searching for the BBAPI
signature will be extremely slow. Additionally, the search will fail
because BBAPI isn't available in a virtual environment. This can be
avoided by skipping the search. Therefore, add a module parameter which
can be used to set the search size to zero.

Signed-off-by: Corvin Köhne <c.koehne@beckhoff.com>
Reviewed-by: Patrick Bruenn <p.bruenn@beckhoff.com>