bbapi: detect Beckhoff BIOS by DMI
bbapi_find_bios will fail on systems without a Beckhoff BIOS because
bbapi isn't available. On Hyper-V bbapi_find_bios takes a huge amount of
time. We introduced a module parameter to avoid it. This was a bad idea
because we have to set the module parameter correctly on each system.
It's much better to read the DMI data to check for a Beckhoff BIOS.

BIOS team told us that all Beckhoff BIOS versions are using the listed
DMI_BIOS_VENDOR strings.

Signed-off-by: Corvin Köhne <c.koehne@beckhoff.com>
Reviewed-by: Steffen Dirkwinkel <s.dirkwinkel@beckhoff.com>
Reviewed-by: Patrick Bruenn <p.bruenn@beckhoff.com>