bbapi: detect Hyper-V by DMI
The scan for bbapi is extremely slow on Hyper-V. We can skip it because
bbapi is only available on Beckhoff hardware with an Beckhoff bios.

The best solution would be to only scan for bbapi when a Beckhoff bios
is detected. Sadly, the DMI data aren't consistent by all Beckhoff bios
versions. So, we use a blacklist of unsupported systems.

Signed-off-by: Corvin Köhne <c.koehne@beckhoff.com>
Reviewed-by: Patrick Bruenn <p.bruenn@beckhoff.com>