bbapi: retry if bbapi_call() fails with BIOSAPI_BUSY
We didn't need it yet but more and more devices are sometimes "busy" and
because the Windows BBAPI driver implemented ten retries with a 100ms
sleep in between, applications don't expect to see that error code and
switch into an error state instead of retrying themself.
So we implement babysitting, too.
Instead of a hardcoded 10 retry limit we add another tunable
busy_retries, which is set to 10 by default, but you can disable retries
and keep the old behaviour by setting this to 0.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>