bbapi: init bios, before we try to write to display
On some devices such as CX20x3 with a BIOS newer than 0.24, the output
on CX2100 display was broken. Well, this was because the BBAPI wasn't
initialized.
It seems smart to initialize the BBAPI in general, before we try to use
it...
A funny thing is that we ignore when BBAPI initialization fails, so we
can now make it a void function.

Fixes: https://azdevops01.beckhoff.com/GSW01/TcOs/_workitems/edit/141043

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>