unittest: make value checking optional
Historically, our bbapi tests were very strict. We had only a few test
devices and a matching config for every test device.

A few years passed and we now have many different test devices and it
isn't possible anymore to define exact expected values for all of these
devices. For this we introduce a similar technic we used with
READ_OBJECT() macro and test_object<..., false> template. We now
introduce the doCompare parameter to all of our check functions. The
default remains to use strict checks, but we can now overwrite the
define UNITTEST_DO_COMPARE with false and would only verify that all the
bbapi calls return the expected return codes.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>