ci: run a minimal unittest on real hardware
In contrast to the past we cannot check for correct value ranges
anymore, as our set of test devices it to large to track all the serial
numbers, boot counters, etc.

So instead we use the new "lazy mode" of our unittest tool and only
verify the correct return codes for our bbapi calls.

Additionally, we use the smalles subset of BBAPI IndexGroups (C6015),
which is a IPC without CX powersupply and display.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>